Add files via upload
This Machine learning module focuses on recommending or predicting the likes of the like minding user by analyzing their sentiments. This model uses sentiment analysis to group the like minded users first then uses a method called Pearson's Correlation. The pre-defined function of Pearson Correlation is not used since it focuses on the calculation of the entire row and not a particular user or variable. So therefore I defined a new function called Pearson's Correlation function where you can find the result between two user depending on their data. Pearson's value ranges from -1 to +1 where -1 is to be extremely unlikely and +1 is to be extremely likely and 0 to be the neutral state. Taking 0.55 as the standard threshold for the like minded users any value about 0.55 would  be considered to be like minded and their suggestions can be predicted depending on the data.